requirements.txt
aiogram>=3.0.0
telethon>=1.25.0
redis>=4.5.0
aioredis>=2.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6 